requirements.txt
lxml==3.4.0
pytz==2014.7
vcrpy==1.1.1
raven==5.0.0
invoke==0.9.0
celery==3.1.15
requests==2.4.1
nameparser==0.3.3
python-dateutil==2.2
fluent-logger==0.3.4
cqlengine==0.21.0
elasticsearch==1.3.0
Flask==0.10.1
blist==1.3.6
furl==0.4.4
jsonschema==2.4.0
jsonpointer==1.7
pycountry==1.10
six==1.9.0